+++ edsnlp/pipes/trainable/embeddings/typing.py
@@ -1,6 +1,7 @@
-from typing import Generic, Optional
+from typing import TYPE_CHECKING, Any, Dict, Generic, List, Optional
 
-import torch
+from foldedtensor import FoldedTensor
+from spacy.tokens import Doc, Span
 from typing_extensions import TypedDict
 
 from edsnlp.core.torch_component import BatchInput, TorchComponent  # noqa: F401
@@ -9,30 +10,62 @@ from edsnlp.utils.span_getters import SpanGetter
 WordEmbeddingBatchOutput = TypedDict(
     "WordEmbeddingBatchOutput",
     {
-        "embeddings": torch.Tensor,
-        "mask": torch.Tensor,
+        "embeddings": FoldedTensor,
     },
 )
 
 
 class WordEmbeddingComponent(
-    Generic[BatchInput], TorchComponent[WordEmbeddingBatchOutput, BatchInput]
+    TorchComponent[WordEmbeddingBatchOutput, BatchInput],
+    Generic[BatchInput],
 ):
     span_getter: Optional[SpanGetter]
     output_size: int
+
+    if TYPE_CHECKING:
+
+        def preprocess(
+            self,
+            doc: Doc,
+            *,
+            contexts: Optional[List[Span]],
+            **kwargs,
+        ) -> Dict[str, Any]:
+            ...
+
+
+class WordContextualizerComponent(
+    Generic[BatchInput], WordEmbeddingComponent[BatchInput]
+):
+    span_getter: Optional[SpanGetter]
+    output_size: int
+    embedding: WordEmbeddingComponent
 
 
 SpanEmbeddingBatchOutput = TypedDict(
     "SpanEmbeddingBatchOutput",
     {
-        "embeddings": torch.Tensor,
-        "mask": torch.Tensor,
+        "embeddings": FoldedTensor,
     },
 )
 
 
 class SpanEmbeddingComponent(
-    Generic[BatchInput], TorchComponent[SpanEmbeddingBatchOutput, BatchInput]
+    TorchComponent[SpanEmbeddingBatchOutput, BatchInput],
+    Generic[BatchInput],
 ):
     span_getter: Optional[SpanGetter]
     output_size: int
+
+    if TYPE_CHECKING:
+
+        def preprocess(
+            self,
+            doc: Doc,
+            *,
+            spans: Optional[List[Span]],
+            contexts: Optional[List[Span]],
+            pre_aligned: bool = False,
+            **kwargs,
+        ) -> Dict[str, Any]:
+            ...