+++ tests/pipelines/trainable/test_transformer.py
@@ -6,6 +6,7 @@ from edsnlp.pipes.trainable.embeddings.transformer.transformer import Transforme
 from edsnlp.pipes.trainable.span_qualifier.span_qualifier import (
     TrainableSpanQualifier,
 )
+from edsnlp.utils.collections import batch_compress_dict, decompress_dict
 
 if not Span.has_extension("label"):
     Span.set_extension("label", default=None)
@@ -44,7 +45,9 @@ def test_span_getter(gold):
         "eds.transformer",
         name="transformer",
         config=dict(
-            model="prajjwal1/bert-tiny", window=128, stride=96, span_getter="to_embed"
+            model="prajjwal1/bert-tiny",
+            window=128,
+            stride=96,
         ),
     )
     nlp.add_pipe(
@@ -54,8 +57,9 @@ def test_span_getter(gold):
             "embedding": {
                 "@factory": "eds.span_pooler",
                 "embedding": nlp.get_pipe("transformer"),
-                "span_getter": ["ents", "sc"],
             },
+            "span_getter": ["ents", "sc"],
+            "context_getter": ["to_embed"],
             "qualifiers": ["_.test_negated", "_.event_type"],
         },
     )
@@ -64,8 +68,20 @@ def test_span_getter(gold):
     qlf.post_init(gold, set())
     batch = qlf.prepare_batch([doc.copy() for doc in gold], supervision=True)
     input_ids = batch["embedding"]["embedding"]["input_ids"]
-    mask = batch["embedding"]["embedding"]["input_mask"]
+    mask = input_ids.mask
     tok = trf.tokenizer
     assert len(input_ids) == 2
     assert tok.decode(input_ids[0][mask[0]]) == "[CLS] arret du ttt si folfox [SEP]"
     assert tok.decode(input_ids[1][mask[1]]) == "[CLS] une autre phrase. [SEP]"
+
+    # Transformer alone with prompts (usually passed by the caller component)
+    prep = trf.preprocess(
+        gold[0],
+        contexts=gold[0].spans["to_embed"],
+        prompts=["Extract the drugs", "Extract the drugs"],
+    )
+    batch = decompress_dict(list(batch_compress_dict([prep])))
+    batch = trf.collate(batch)
+    batch = trf.batch_to_device(batch, device=trf.device)
+    res = trf(batch)
+    assert res["embeddings"].shape == (2, 5, 128)